+++ packages/core/src/mdx-plugins/index.ts
@@ -4,7 +4,5 @@ export {
 } from 'remark-gfm';
 export * from './rehype-code';
 export * from './remark-image';
-export * from './remark-dynamic-content';
 export * from './remark-structure';
 export * from './remark-heading';
-export * from './remark-install';