+++ packages/core/src/source/path.ts
@@ -23,8 +23,8 @@ export interface FileInfo {
 }
 
 export function parseFilePath(path: string): FileInfo {
-  const slashedPath = slash(path);
-  const parsed = parse(slashedPath);
+  const normalized = normalizePath(path);
+  const parsed = parse(normalized);
   const flattenedPath = joinPaths([parsed.dir, parsed.name]);
   const [name, locale] = parsed.name.split('.');
 
@@ -33,22 +33,34 @@ export function parseFilePath(path: string): FileInfo {
     name,
     flattenedPath,
     locale,
-    path: slashedPath,
+    path: normalized,
   };
 }
 
 export function parseFolderPath(path: string): FileInfo {
-  const slashedPath = slash(path);
-  const parsed = parse(slashedPath);
+  const normalized = normalizePath(path);
+  const parsed = parse(normalized);
   const [name, locale] = parsed.base.split('.');
 
   return {
     dirname: parsed.dir,
     name,
-    flattenedPath: slashedPath,
+    flattenedPath: normalized,
     locale,
-    path: slashedPath,
+    path: normalized,
   };
+}
+
+/**
+ * @param path - Relative path
+ * @returns Normalized path, with no trailing/leading slashes
+ * @throws Throws error if path starts with `./` or `../`
+ */
+export function normalizePath(path: string): string {
+  const segments = splitPath(slash(path));
+  if (segments[0] === '.' || segments[0] === '..')
+    throw new Error("It must not start with './' or '../'");
+  return joinPaths(segments);
 }
 
 /**