+++ packages/typescript/src/index.ts
@@ -1,3 +1,4 @@
 export * from './generate/base';
 export * from './generate/mdx';
 export * from './generate/file';
+export { renderMarkdownToHast } from './markdown';